A different way to do config for providers
My main goal here was to let it be more flexible to keep
a nicer UX for hand-editing the config file; we want
providers to be able to load values from different
places in the config file.

As example cases, the following now work.

In project.yml, you can set defaults:

  runtime:
    REDIS_URL:
      default: "redis://example.com"

In .anaconda/project-local.yml you can set local values for variables:

  variables:
    REDIS_URL: "redis://example.com"

Also providers can have more esoteric config, the example I used was
the port range for ProjectScopedRedisProvider, like this:

  runtime:
    REDIS_URL:
      providers:
        ProjectScopedRedisProvider:
          port_range: 7000-7100

In this commit, "default" for a runtime requirement is ignored in
project-local.yml and "providers" is ignored in project.yml, maybe
later we'll decide to allow either one in either place though.
Not sure yet.