Implement simple configuration of providers via the HTML UX
 - a provider can provide an HTML form snippet
 - we insert this snippet into an exceptionally ugly
   and user-unfriendly HTML page
 - we save any changes made via the snippet to project-local.yml

This commit also adds a "prepare" command which opens the browser
UI and then performs the prepare().

I'm pretty sure some of this isn't right (we need the HTML UX to
be closer to something real, and need more examples of providers,
to really work it out). But it's a start we can iterate.

Fixes old issue 1 (the command that opens the UI)
Fixes old issue 6 (form lets you customize Redis provider)
Fixes old issue 7 (form customizations are saved to local state file)