Safeguard YAML files against overwriting due to corruption
We don't want to accidentally overwrite a YAML file just because
someone put a syntax error in there.

Add a `corrupted` flag which can be used by a UI to display the
problem; if a program ignores this flag, then attempts to modify
the YAML file will raise an exception rather than overwriting
the file.