Introduce RequirementStatus and Requirement.check_status
This cleans up the various `requirements_and_providers` code and
also allows us to make status display arbitrarily intricate.
As a starting point, we can now display details of a "green"
status for a requirement.

RequirementStatus also acts as a cache for any IO that the previous
why_not_provided method would do.

This commit improves some of the status descriptions from Conda env, redis,
and master password requuirements, too.

This commit also slightly improves the HTML UI to show the status
in addition to the name of the requirement.