Every requirement now has exactly 1 provider, found by class name
The one provider could chain to many ways to provide, or possibly
even multiplex sub-providers, but we shouldn't try to deal with
multiple in generic fashion across all the code. It was too unclear
how that should work.

We may want to look up providers in some way other than class name
later, or we may want to switch to simply importing providers
rather than having a lookup (so the plugin system would only find
requirements, and requirements would come with providers).