Use a minimal rather than empty environment in unit tests
This fixes old issue 68 than on most systems we do need the PATH set
for the tests to pass. Some other variables are also included
on the basis that they are likely to be important and should not
break our tests.

Potentially we should blacklist rather than whitelist system env
vars, but let's try this for a while and see if it's fine.