Support multiple named commands rather than single app:entry:
Commands now go under a `commands` section which is a dict from
command names to command attributes.

This keeps the command line format the same, and calls it
`conda_app_entry` - which is sort of an odd thing, but
trying to avoid tackling multiple problems in one commit.
We could drop the `conda_app_entry` support once we add
other command line formats in a subsequent commit, or we
could keep it to tunnel through meta.yaml's app:entry:
as we do in this commit.

Fixes old issue 64