Adjust how we get Project.name and remove `version`
meta.yaml compat isn't as important as we were thinking. Use
toplevel `name:` instead of `package: name:` in project.yml,
fall back to name from meta.yaml and then to directory name.
Report errors if the name isn't a string.

Remove all references to `version` because we don't have any
actual use for that right now; we can add it back if/when we do.