Modify CondaEnvRequirement/Provider to support CONDA_DEFAULT_ENV on Windows
We have a problem that on Windows, only CONDA_DEFAULT_ENV is set, and on
Unix both CONDA_DEFAULT_ENV and CONDA_ENV_PATH are set, but CONDA_DEFAULT_ENV
may be an env name instead of an env prefix.

This commit changes CondaEnvRequirement so that the env var it uses is
always fixed by the platform, and it can no longer appear in the runtime
section (there was no reason to put it there anyway).

This commit passes tests on Linux but hasn't been tested on Windows yet
and probably doesn't work; however it lays the groundwork for Windows
to have a shot at working. Will iterate it further over on Windows.