Add `anaconda-project init` subcommand
Fixes old issue 50

For now the project directory is provided the same as for
the other subcommands, with --project or defaulting to
current directory. It does feel a bit weird that you can't
say `anaconda-project init mydir` though, so maybe we'll
have to add that later as well as a synonym for --project.