Implement `anaconda-project add-download`
A shortcoming of this commit is that if we can't download the URL,
we don't display the actual problem, just a generic error message.
To fix this we'll have to get the errors from provide() attached to
specific RequirementStatus instances, but don't want to mix that
in to this commit.

Another limitation is that we don't offer a way to set the filename
or the content hash.

Fixes old issue 133, with those caveats.