Write a version.py from setup.py to avoid coverage bugs
The problem with the previous approach is that setup.py
imported anaconda_project before the coverage stuff
started to track coverage, so we didn't get coverage stats
on the anaconda_project module. This problem could be
nocov'd as a hack, but other problems like that could happen
too perhaps.

This commit changes things to avoid making the build depend
on the code, which should be more robust.