For Redis options in project-local.yml, use service_options not runtime
Also considered putting the options in `variables` as in project.yml,
but since we have another issue to move Redis to `services` in project.yml,
that seems shortsighted.

We may change this later but since I'm about to get rid of the runtime
section in project.yml wanted to get rid of it here too.