Remove need to preflight-check project.problems for most project_ops
Now that we have the Status class, we can use it to return a failed
status containing the project.problems, so callers don't have to
special-case handle project.problems. Callers simply deal with the
returned Status.

This cleans up our code in the command line tool and should also clean
up code in other users of the API.