Fix CondaEnvRequirement if the desired env exists but isn't set in the env
If you already created projectdir/envs/default, but it isn't set as
CONDA_ENV_PATH, then CondaEnvRequirement would claim that the requirement
was met (the env exists and CONDA_ENV_PATH was set). We need to include
the value of CONDA_ENV_PATH in what we check for status.

Fixes old issue 188