Don't auto-detect notebooks in envs/ or in dot-dirs
dot-dirs could have the git cache, egg packages, etc.
envs/ has whole environments.

If nothing else it takes a long time to walk these large
directories, so we want to ignore them for that reason,
but the immediate reason for this patch is that envs
commonly do have .ipynb files in them which should
not become commands.

Fixes old issue 207