When removing dependencies from a specific env, also remove from global
If we ask to remove dependency 'foo' from env 'bar', we need to
also remove 'foo' from the global dependencies section so 'bar'
won't inherit it, and then re-add 'foo' to any non-'bar' envs
so that their package set is unchanged.

The net result is supposed to be that if you delete a package
it in fact goes away.

Fixes old issue 234