Start prepare() process with conda env unset
Previously we special-cased conda-env-related variables so that
they could be set (but we'd ignore them) during prepare. Now,
we initially clear them out, and pass the previous value through
the UserConfigOverrides. This required passing the "overrides" around
to several more places.

The reason for the change is to enable us to tell whether the
provide() for the conda env variables has run yet, so we can
require it to run prior to performing later steps. If we make
any or all other providers require an environment, then we need
to know whether we've already picked the environment.