Remove master password encryption in favor of using OS keyring
This should be a little more intuitive for people, hopefully.
I'm not sure how the UX plays out on all platforms though.
We'll have to experiment.

The browser-based setup now always starts with establishing an environment,
and only then can move on to everything else.
This is because we associate the stored passwords with the environment...
the idea is that we probably want to move all of our state to be
environment-associated. Right now most other state is project-directory-associated
though, so it's a little strange for now.

Dropped _ENCRYPTED from the secret suffixes list because it seems a little weird.

Re-enabled encrypted variables by default.

Removed bcrypt and pycrypto dependencies.