Add some force-resets of the keyring global state to tests
This may fix test failures caused by tests that leave global
state modified, but won't fix failures caused by tests running
in parallel in the same process. To fix those we would need to
make the keyring module use a class to avoid global state, and pass
an instance of that class around as appropriate.