.travis.yml unshallow git before build; only build packages on 3.5
We need an unshallow clone in order to get our build number and
tag information, and because conda requires it to clone from
the clone.

We build all arches at once using "conda convert" so we don't
need to make packages for the whole matrix, only one of the
elements in the matrix.