Make a non-empty env_specs section mandatory
This is the first step in moving to avoid "implicit"
kapsel configuration; the envs and commands that exist
instead will have to be explicitly in kapsel.yml.
The purpose of this is clarity. The downside is that
an empty kapsel.yml is no longer valid.

Tons of tests had to be fixed to contain explicit env_specs
sections.