Add generic http server configuration options to commands
The idea here is that you can start an http server command
without special-casing Bokeh, Notebooks, or other kinds of command.

Also, bokeh and notebook commands will now both default to opening
a browser to the new http server (previously Bokeh did not).

`conda kapsel run` now supports options
--kapsel-host, --kapsel-port, --kapsel-url-prefix, and --kapsel-no-browser.
These are auto-translated into the appropriate options for Bokeh and notebooks,
and will be passed through unchanged from `conda kapsel run` to other commands.

A new command attribute `supports_http_options` is available,
which defaults to true for Bokeh and notebooks, and has to be set to true
manually for shell commands. For shell commands the http options are passed
through verbatim regardless of this attribute, but setting it to true
allows other code to ask whether a command can be run as an http server.

For bokeh and notebooks, setting it to false disables translation of
the --kapsel-* options, which probably isn't useful, but just being
complete. Maybe it's useful to disable if you don't want kapsel to
auto-add `--show` to a bokeh server or something.