Update timestamp file always even when not making changes to the env
Otherwise, we keep having to re-analyze the environment even though
nothing has changed.

The bulk of this commit is rearranging things so we call
fix_environment_deviations for an "inherited" environment, which
comes up in the tests a lot. This does mean we will now modify
inherited environments (set via CONDA_PREFIX) including adding
our timestamp files and installing packages.