Avoid getLogger because it confuses pytest
The issue (I think...) is that the StreamHandler is created with a pytest
mock sys.stderr that gets closed and then a later test will get a bad
handler. Global variables yay!

So create our Logger directly without using getLogger. We are really using
logger as a sort of implementation detail anyhow to avoid creating our
own similar interface.