When running a notebook, set the default_url to that notebook
This changes browsing to / into /notebooks/foo.ipynb instead of /tree.

In the default case where we open a browser, this makes little difference
because Jupyter opens to /notebooks/foo.ipynb instead of / anyhow.
It does break where clicking the Jupyter logo takes you; wish that
was hardcoded to /tree in Jupyter.

The purpose of this change is when we're deploying a notebook as an http
app, / does the right thing. Potentially we should only set default_url
when --kapsel-no-browser is provided? But that seems sort of magic, so
not doing that for now.