Merge in global packages/channels using env spec inheritance
The reason this matters: otherwise each env spec includes
and "reiterates" the global list, which means the global
list gets output in EnvSpec.to_json, and also that it
overrides package versions specified in parent env specs.
Instead, we'd like the global lists to be the "root" and
all intermediate ancestors should win over it.