Add is_dict and is_list utilities and port a few places to use them
The main purpose is to allow non-dict/list objects to be treated as
such, so that YamlFile could have proxy wrappers around its tree
of nodes.

This is by no means all the isinstance cases that should be ported,
decided not to do all that work in the current branch. Can come back
later and use these in more places.