Revert/save both project file and lock file in ops "transaction"
The unfortunate side effect of this for now is that even if
only one file was modified, we resave both, possibly causing
some whitespace changes and mtime update. The fix will be to
track the YamlFile._dirty flag more accurately by either
proxying all the dicts/lists in the yaml, or requiring anyone
who modifies them directly to call something like YamlFile.mark_dirty().
Don't want to mix that into this branch though; let's live with
the extra file rewrites and fix it as a separate bug.