When adding/removing packages/env_specs, update the lock file
At present, if you add/remove anything we re-resolve dependencies
which means we update them all to latest. It might be nice if
we did some sort of "minimal required" update, but that's a
more elaborate (and slower for the user) thing to implement.

Another limitation of this commit is that we will create an
anaconda-project-lock.yml when you didn't have one before,
if you add/remove anything. Will fix that in a later commit.