Include "default" for variables in publication_info()
If the default is None/missing we leave it out of the JSON.

We canonicalize the default to a string, even if it was an int
or float in anaconda-project.yml, so the server side doesn't have
to deal with that.

Also did a little DRY cleanup in publication_info() across the
requirement types. This leads to supplying encrypted=False for
downloads and services, but that seems fine.