Remove browser UI mode.
In this PR, keeping around read_config/set_config_values_as_strings
even though only the HTML UI really used those, because in theory
we might want to use them to make a similar textual UI (or an HTML
UI that isn't included in the anaconda-project codebase).

This commit can be referenced as a guide to building a textual
UI in the future.