Drop "buffered" logs from status objects, use streaming to Frontend
Still keeping a copy of the "errors" buffer in the status objects,
since it might be nice to show all of those in a dialog for example.
This could be a poor decision, not sure.

This breaks the API in some ways, not sure if it will break any actual
third party code or not. Not sure what code is out there.