Add env_spec_name parameter to variable-affecting commands
This adds env_spec_name param to project_ops for variables,
and a --env-spec option to the command line subcommands.

There's a trick that variables are actually per-env-prefix
(well, keyring vars are and local state file vars aren't, but
that's something of a bug). This means we need to prepare to
see what the env prefix is. Adopted this policy:
 - can pass in a prepare_result when doing set/unset/remove
 - use env prefix from prepare result when provided
 - otherwise, do a prepare to get env prefix. For set, do the
   prepare in dev mode so we will create the env if missing;
   for unset, use check mode and error if not already created;
   for remove, use check mode and silently don't unset if
   not already created.

This approach needs to be copied to downloads and services, next.