TST: Updated test_prepare.py::test_default_to_system_environ for CONDA_STACK_2.
If the user has `auto_stack: 1` or higher and has a stacked environment
as indicated by `CONDA_STACK_2 == True`, then the PATH environmental
variable might be modified bt prepare.  This change ignores the error if
`CONDA_STACK_2 == True`.  A better resolution would be to control the
configuration of the conda environment before running tests.