mock logical hash when default platforms are injected
some tests intentionall do not set platforms in the project yaml.
If the test is run on a non-default platform, like osx-arm64
the hash is different because the platforms list
is different. A mocked logical_hash is used when
necessary to remove the platforms so that the
test can pass on all hardware.